Support default exports
Currently the plugin doesn't handle `{import('./module').default}` syntax correctly if it's specified as type, despite that VS Code handles it well. With the change, the name of the module will be used as the name of the type in JSDoc (a typical use case when classes are exported). The change is not expected to be breaking: `default` is a reserved word, and exports can't be named like that.